Add files via upload
I'm excited to share that I've finished my first project as an intern at CodSoft! The goal of the project was to predict the survival of passengers on the Titanic using a machine learning model called logistic regression.



I'm happy to report that I was able to complete the project with a 96.39% accuracy. This means that my model was able to correctly predict the survival of 96.39% of the passengers in the Titanic dataset.



I'm really proud of what I was able to accomplish with this project. I learned a lot about machine learning and data science, and I'm excited to continue learning and growing in this field.



I'm grateful for the opportunity to have worked on this project, and I'm thankful for the support of my mentors and colleagues at CodSoft.





colab code: